Adding more details to README.md + login for phpunit generated

+++ src/PHPProjectGen.php
@@ -59,7 +59,7 @@ class PHPProjectGen
         $composer_data['license'] = $this->composer_config['license'];
         $composer_data['authors'][0]['name'] = $this->composer_config['developer'];
         $composer_data['authors'][0]['email'] = $this->composer_config['email'];
-        $src = $this->composer_config['name'] . "\\\\" . $this->composer_config['projectname']."\\\\";
+        $src = $this->composer_config['name'] . "\\\\" . $this->composer_config['projectname'] . "\\\\";
         $tests = $this->composer_config['name'] . "\\\\" . $this->composer_config['projectname'] . "\\\\Test\\\\";
         $composer_data['autoload']['psr-0'] = [
           $src => "src/",
@@ -113,17 +113,27 @@ class PHPProjectGen
         $zipFile = new \PhpZip\ZipFile();
         $zipFile
           ->addFile(__DIR__ . "/temp/" . $this->composer_config['projectname'] . ".php", "src/" . $this->composer_config['projectname'] . ".php")
-          ->addFile(__DIR__ . "/temp/test" . $this->composer_config['projectname'] . ".php", "tests/test" . $this->composer_config['projectname'] . ".php")
           ->addFile(__DIR__ . "/temp/composer.json", "composer.json")
           ->addFile(__DIR__ . "/temp/.gitignore", ".gitignore")
-          ->addFromString("README.md", "#" . $this->composer_config['projectname'])
-          ->saveAsFile($this->composer_config['projectname'].'.zip')
+          ->addFromString("README.md", "#" . $this->composer_config['projectname']);
+
+        if ($this->composer_config['phpunit']) {
+            $zipFile
+              ->addFile(__DIR__ . "/temp/test" . $this->composer_config['projectname'] . ".php", "tests/test" . $this->composer_config['projectname'] . ".php");
+        }
+
+        $zipFile
+          ->saveAsFile($this->composer_config['projectname'] . '.zip')
           ->close();
+
     }
 
-    private function cleanData(){
+    private function cleanData()
+    {
         unlink(__DIR__ . "/temp/" . $this->composer_config['projectname'] . ".php");
-        unlink(__DIR__ . "/temp/test" . $this->composer_config['projectname'] . ".php");
+        if ($this->composer_config['phpunit']) {
+            unlink(__DIR__ . "/temp/test" . $this->composer_config['projectname'] . ".php");
+        }
         unlink(__DIR__ . "/temp/composer.json");
     }
